refactor: move some modules from ui-utils
