+++ apps/meteor/client/components/message/MessageToolbarHolder.tsx
@@ -3,7 +3,7 @@ import { MessageToolbarWrapper } from '@rocket.chat/fuselage';
 import { useQuery } from '@tanstack/react-query';
 import { Suspense, lazy, memo, useState } from 'react';
 
-import type { MessageActionContext } from '../../../app/ui-utils/client/lib/MessageAction';
+import type { MessageActionContext } from '../../lib/MessageAction';
 import { useChat } from '../../views/room/contexts/ChatContext';
 import { useIsVisible } from '../../views/room/hooks/useIsVisible';
