+++ apps/meteor/client/components/message/toolbar/useMarkAsUnreadMessageAction.ts
@@ -2,7 +2,7 @@ import { isOmnichannelRoom } from '@rocket.chat/core-typings';
 import type { ISubscription, IMessage, IRoom } from '@rocket.chat/core-typings';
 import { useRouter, useUser } from '@rocket.chat/ui-contexts';
 
-import type { MessageActionConfig } from '../../../../app/ui-utils/client/lib/MessageAction';
+import type { MessageActionConfig } from '../../../lib/MessageAction';
 import { useMarkAsUnreadMutation } from '../hooks/useMarkAsUnreadMutation';
 
 export const useMarkAsUnreadMessageAction = (
