+++ apps/meteor/client/components/message/toolbar/usePinMessageAction.tsx
@@ -2,7 +2,7 @@ import type { IMessage, IRoom, ISubscription } from '@rocket.chat/core-typings';
 import { isOmnichannelRoom } from '@rocket.chat/core-typings';
 import { useSetting, useSetModal, usePermission } from '@rocket.chat/ui-contexts';
 
-import type { MessageActionConfig } from '../../../../app/ui-utils/client/lib/MessageAction';
+import type { MessageActionConfig } from '../../../lib/MessageAction';
 import PinMessageModal from '../../../views/room/modals/PinMessageModal';
 import { usePinMessageMutation } from '../hooks/usePinMessageMutation';
