+++ apps/meteor/client/components/message/toolbar/useReportMessageAction.tsx
@@ -1,7 +1,7 @@
 import type { ISubscription, IRoom, IMessage } from '@rocket.chat/core-typings';
 import { useSetModal, useUser } from '@rocket.chat/ui-contexts';
 
-import type { MessageActionConfig } from '../../../../app/ui-utils/client/lib/MessageAction';
+import type { MessageActionConfig } from '../../../lib/MessageAction';
 import { roomCoordinator } from '../../../lib/rooms/roomCoordinator';
 import ReportMessageModal from '../../../views/room/modals/ReportMessageModal';
