+++ apps/meteor/client/components/message/toolbar/useTranslateAction.ts
@@ -2,7 +2,7 @@ import type { IMessage, ISubscription, IRoom } from '@rocket.chat/core-typings';
 import { useEndpoint, usePermission, useSetting, useUser } from '@rocket.chat/ui-contexts';
 import { useMemo } from 'react';
 
-import type { MessageActionConfig } from '../../../../app/ui-utils/client/lib/MessageAction';
+import type { MessageActionConfig } from '../../../lib/MessageAction';
 import { AutoTranslate } from '../../../lib/autotranslate';
 import { roomCoordinator } from '../../../lib/rooms/roomCoordinator';
 import { Messages } from '../../../stores';
